Fix dfu controller not rediscovered when dfu fails
The nordic dfu sdk resets delegates on the peripheral. This way, we don’t realize that the peripheral has during flashing or when turned off. As the user can restart the dfu process we also want to re-discover the same peripheral. That’s why we simply reset the dfu discovery manager and thus we forget any dfu controller that had been found before.